pyproject.toml
[project]
name = "haruka-gallery"
version = "0.1.0"
description = "Nonebot plugin Haruka Gallery"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "image>=1.5.33",
    "imagehash>=4.3.2",
    "numpy>=2.3.4",
    "pilmoji>=2.0.4",
    "setproctitle>=1.3.7",
    "nonebot2>=2.4.4,<3.0.0",
    "nonebot-adapter-onebot>=2.4.6",
]

[tool.nonebot]
adapters = [
    { "name" = "Onebot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = ["haruka_gallery"]
